HHH-17884 Fix the enum value sorting for EnumeratedType.STRING enums
