+++ hibernate-core/src/main/java/org/hibernate/dialect/MySQLEnumJdbcType.java
@@ -6,22 +6,7 @@
  */
 package org.hibernate.dialect;
 
-import org.hibernate.type.descriptor.ValueBinder;
-import org.hibernate.type.descriptor.ValueExtractor;
-import org.hibernate.type.descriptor.WrapperOptions;
-import org.hibernate.type.descriptor.java.JavaType;
-import org.hibernate.type.descriptor.jdbc.BasicBinder;
-import org.hibernate.type.descriptor.jdbc.BasicExtractor;
-import org.hibernate.type.descriptor.jdbc.JdbcLiteralFormatter;
-import org.hibernate.type.descriptor.jdbc.JdbcType;
-import org.hibernate.type.descriptor.jdbc.internal.JdbcLiteralFormatterCharacterData;
-
-import java.sql.CallableStatement;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import static org.hibernate.type.SqlTypes.ENUM;
+import org.hibernate.type.descriptor.jdbc.EnumJdbcType;
 
 /**
  * Represents an {@code enum} type on MySQL.
@@ -31,60 +16,11 @@ import static org.hibernate.type.SqlTypes.ENUM;
  *
  * @see org.hibernate.type.SqlTypes#ENUM
  * @see MySQLDialect#getEnumTypeDeclaration(String, String[])
+ * @deprecated Use {@link EnumJdbcType} instead
  *
  * @author Gavin King
  */
-public class MySQLEnumJdbcType implements JdbcType {
+@Deprecated
+public class MySQLEnumJdbcType extends EnumJdbcType {
 
-	@Override
-	public int getJdbcTypeCode() {
-		return ENUM;
-	}
-
-	@Override
-	public <T> JdbcLiteralFormatter<T> getJdbcLiteralFormatter(JavaType<T> javaType) {
-		return new JdbcLiteralFormatterCharacterData<>( javaType );
-	}
-
-	@Override
-	public String getFriendlyName() {
-		return "ENUM";
-	}
-
-	@Override
-	public <X> ValueBinder<X> getBinder(JavaType<X> javaType) {
-		return new BasicBinder<>( javaType, this ) {
-			@Override
-			protected void doBind(PreparedStatement st, X value, int index, WrapperOptions options)
-					throws SQLException {
-				st.setString( index, getJavaType().unwrap( value, String.class, options ) );
-			}
-
-			@Override
-			protected void doBind(CallableStatement st, X value, String name, WrapperOptions options)
-					throws SQLException {
-				st.setString( name, getJavaType().unwrap( value, String.class, options ) );
-			}
-		};
-	}
-
-	@Override
-	public <X> ValueExtractor<X> getExtractor(JavaType<X> javaType) {
-		return new BasicExtractor<>( javaType, this ) {
-			@Override
-			protected X doExtract(ResultSet rs, int paramIndex, WrapperOptions options) throws SQLException {
-				return getJavaType().wrap( rs.getString( paramIndex ), options );
-			}
-
-			@Override
-			protected X doExtract(CallableStatement statement, int index, WrapperOptions options) throws SQLException {
-				return getJavaType().wrap( statement.getString( index ), options );
-			}
-
-			@Override
-			protected X doExtract(CallableStatement statement, String name, WrapperOptions options) throws SQLException {
-				return getJavaType().wrap( statement.getString( name ), options );
-			}
-		};
-	}
 }
